cleanup response trait
